+++ autogpt_platform/frontend/src/components/node-input-components.tsx
